🐛 fix(sal): DwtGetTimeline_us 长时运行 uint32 溢出
dwt_time_.s * 1000000 为 uint32 乘法, 连续运行约 71.6 分钟 (s >= 4295)
后返回值回绕。先提升到 64 位再乘。Phase B 测试基建首批用例发现,
经用户确认不关联飞书单号。

回归用例: DwtMath.TimelineUsSurvivesLongUptime (200 次溢出 ≈ 5113 s)

+++ Core/sal/dwt/sal_dwt.cc
@@ -84,7 +84,8 @@ float DwtInstance::DwtGetTimeline_ms(void)
 uint64_t DwtInstance::DwtGetTimeline_us(void)
 {
     DwtSysTimeUpdate();
-    return dwt_time_.s * 1000000 + dwt_time_.ms * 1000 + dwt_time_.us;
+    // s 必须先提升到 64 位再乘: uint32 乘法在 s >= 4295 (约 71.6 分钟) 时回绕
+    return (uint64_t)dwt_time_.s * 1000000 + (uint64_t)dwt_time_.ms * 1000 + dwt_time_.us;
 }
 
 void DwtInstance::DwtDelay(float Delay)
